Fixes floating point output in decorators

+++ header/SpikesDecorator.hpp
@@ -11,7 +11,7 @@ class SpikesDecorator : public AbilityDecorator {
     virtual int defend(vector<BaseCharacter*> charList, int attackerIndex, int damage) {
         if (rand() % 2) {
             charList.at(attackerIndex)->defend(charList, attackerIndex, (attackStrength * (0.0236 + sqrt(0.02512 * this->getLevel() + 0.00573168)) / 0.01256) / 100);
-            addAbilityMessage(charList.at(attackerIndex)->getName() + " was thorned for " + to_string((attackStrength * 0.186 + 0.102 * this->getLevel() + 6.28 * pow(10, -3) * pow(this->getLevel(), 2) + 10) / 100) + " damage whilst trying to hurt " + this->character->getName());
+            addAbilityMessage(charList.at(attackerIndex)->getName() + " was thorned for " + to_string(static_cast<int>((attackStrength * 0.186 + 0.102 * this->getLevel() + 6.28 * pow(10, -3) * pow(this->getLevel(), 2) + 10) / 100)) + " damage whilst trying to hurt " + this->character->getName());
             cout << this->currentMessage << endl;
             this->currentMessage = "";
         }
